Outsource range handling to separate file
